updated plotbands for batt_volt and inverter gauges.
Now battery voltage has a red band at the top as well as the bottom.
Inverters now are set to show the “30 min rated” power output as being
in the red. I consider the 5 second and surge values to be “over the
top” and would rather not see them in my gauge thus lowering my
precision.